% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/age_hist1.R
\name{age_hist1}
\alias{age_hist1}
\title{Distribution of ages of professors at Williams College}
\usage{
age_hist1(x)
}
\arguments{
\item{\code{x}}{An integer on [2004, 2014]}
}
\description{
Returns a histogram of ages of professors at Williams College for any year from 2004 to 2014.
}
\details{
For this to work properly, argument must be an integer representing a year from 2004 to 2014.
}

