% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ages_1.R
\name{ages_1}
\alias{ages_1}
\title{Age vector generator 1}
\usage{
ages_1(x)
}
\arguments{
\item{\code{x}}{An integer on [2004, 2013]}
}
\description{
Generates a vector filled with integers representing the ages of professors/lecturers at Williams College in any year from 2004 to 2013.
}
\details{
For this to work properly, argument must be an integer representing a year from 2004 to 2013.
}

