% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ages_2.R
\name{ages_2}
\alias{ages_2}
\title{Age vector generator 2}
\usage{
ages_2(x)
}
\arguments{
\item{\code{x}}{the integer 2014}
}
\description{
Generates a vector filled with integers representing the ages of professors/lecturers at Williams College in 2014.
}
\details{
For this to work properly, argument must be the integer 2014.
}

