% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/avg_age.R
\name{avg_age}
\alias{avg_age}
\title{Average age of professors at Williams College.}
\usage{
avg_age(x)
}
\arguments{
\item{\code{x}}{An integer on [2004, 2014]}
}
\description{
Returns the average age of professors at Williams College in any year from 2004 to 2014.
}
\details{
For this to work properly, argument must be an integer representing a year from 2004 to 2014.
}

