% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/yngst.R
\name{yngst}
\alias{yngst}
\title{Youngest professor at Williams College}
\usage{
yngst(x)
}
\arguments{
\item{\code{x}}{An integer on [2004, 2014]}
}
\description{
Returns the name(s) and age(s) of the youngest professor(s) at Williams college in a year from 2004 to 2014.
}
\details{
For this to work properly, argument must be an integer representing a year from 2004 to 2014.
}

